refactor: update backend to be more structured on the admin side
