+++ services/admin.py
@@ -33,177 +33,6 @@ ORDER BY month;
         db.session.rollback()
         return {"error": str(e)}
 
-def get_image_detection_accuracy():
-    try:
-        query = text("""
-SELECT 
-    TO_CHAR(date_of_guess, 'YYYY-MM') AS month,
-    SUM(CASE WHEN user_guess_type = (SELECT image_type FROM images WHERE images.image_id = user_guesses.image_id) THEN 1 ELSE 0 END) * 1.0 / COUNT(*) AS accuracy
-FROM user_guesses
-WHERE date_of_guess >= NOW() - INTERVAL '12 months'
-GROUP BY month
-ORDER BY month;
-        """)
-        result = db.session.execute(query)
-        db.session.commit()
-
-        rows = []
-        for row in result:
-            row_dict = {column: value for column, value in zip(result.keys(), row)}
-            rows.append(row_dict)
-
-        return rows
-    except Exception as e:
-        db.session.rollback()
-        return {"error": str(e)}
-    
-def get_confusion_matrix():
-    try:
-        query = text("""
-            SELECT
-                SUM(CASE WHEN user_guess_type = 'real' AND image_type = 'real' THEN 1 ELSE 0 END) AS truePositive,
-                SUM(CASE WHEN user_guess_type = 'ai' AND image_type = 'real' THEN 1 ELSE 0 END) AS falsePositive,
-                SUM(CASE WHEN user_guess_type = 'real' AND image_type = 'ai' THEN 1 ELSE 0 END) AS falseNegative,
-                SUM(CASE WHEN user_guess_type = 'ai' AND image_type = 'ai' THEN 1 ELSE 0 END) AS trueNegative
-            FROM user_guesses
-            JOIN images ON user_guesses.image_id = images.image_id;
-        """)
-        result = db.session.execute(query)
-        db.session.commit()
-
-        confusion_matrix = {}
-        for row in result:
-            confusion_matrix = {column: value for column, value in zip(result.keys(), row)}
-
-        return confusion_matrix
-    except Exception as e:
-        db.session.rollback()
-        return {"error": str(e)}
-    
-    
-def get_image_confusion_matrix(image_id):
-    try:
-        query = text(f"""
-            SELECT
-                SUM(CASE WHEN user_guesses.user_guess_type = 'real' AND images.image_type = 'real' THEN 1 ELSE 0 END) AS truePositive,
-                SUM(CASE WHEN user_guesses.user_guess_type = 'ai' AND images.image_type = 'real' THEN 1 ELSE 0 END) AS falsePositive,
-                SUM(CASE WHEN user_guesses.user_guess_type = 'real' AND images.image_type = 'ai' THEN 1 ELSE 0 END) AS falseNegative,
-                SUM(CASE WHEN user_guesses.user_guess_type = 'ai' AND images.image_type = 'ai' THEN 1 ELSE 0 END) AS trueNegative
-            FROM user_guesses
-            JOIN images ON user_guesses.image_id = images.image_id
-            WHERE images.image_id = {image_id}
-        """)
-        
-        result = db.session.execute(query)
-        db.session.commit()
-
-        confusion_matrix = {}
-        
-        for row in result:
-            confusion_matrix = {column: value for column, value in zip(result.keys(), row)}
-
-        return confusion_matrix
-    
-    except Exception as e:
-        db.session.rollback()
-        return {"error": str(e)}
-
-
-def get_leaderboard():
-    try:
-        query = text("""
-SELECT 
-    user_guesses.user_id, 
-    username,
-    AVG(CASE WHEN user_guess_type = image_type THEN 1 ELSE 0 END) AS accuracy
-FROM user_guesses
-JOIN images ON user_guesses.image_id = images.image_id
-JOIN users ON users.user_id = user_guesses.user_id
-GROUP BY user_guesses.user_id, users.username
-ORDER BY accuracy DESC
-LIMIT 10;
-        """)
-        result = db.session.execute(query)
-        db.session.commit()
-
-        leaderboard = []
-        for row in result:
-            leaderboard.append({column: value for column, value in zip(result.keys(), row)})
-
-        return leaderboard
-    except Exception as e:
-        db.session.rollback()
-        return {"error": str(e)}
-
-def get_image_difficulty():
-    try:
-        query = text("""
-            SELECT 
-                images.image_id, 
-                images.image_path,
-                COUNT(*) AS total_guesses,
-                SUM(CASE WHEN user_guess_type != images.image_type THEN 1 ELSE 0 END) AS incorrect_guesses,
-                (SUM(CASE WHEN user_guess_type != images.image_type THEN 1 ELSE 0 END) * 1.0 / COUNT(*)) AS difficulty_score
-            FROM user_guesses
-            JOIN images ON user_guesses.image_id = images.image_id
-            GROUP BY images.image_id
-            ORDER BY difficulty_score DESC;
-        """)
-        result = db.session.execute(query)
-        db.session.commit()
-
-        difficulty_data = []
-        for row in result:
-            difficulty_data.append({column: value for column, value in zip(result.keys(), row)})
-
-        return difficulty_data
-    except Exception as e:
-        db.session.rollback()
-        return {"error": str(e)}
-
-def get_ml_metrics():
-    try:
-        # Use text() to ensure that SQLAlchemy knows it's a raw SQL query
-        query = text("""
-            SELECT 
-                SUM(CASE WHEN user_guesses.user_guess_type = images.image_type AND user_guesses.user_guess_type = 'real' THEN 1 ELSE 0 END) AS true_positive,
-                SUM(CASE WHEN user_guesses.user_guess_type != images.image_type AND user_guesses.user_guess_type = 'real' THEN 1 ELSE 0 END) AS false_positive,
-                SUM(CASE WHEN user_guesses.user_guess_type = images.image_type AND user_guesses.user_guess_type = 'ai' THEN 1 ELSE 0 END) AS true_negative,
-                SUM(CASE WHEN user_guesses.user_guess_type != images.image_type AND user_guesses.user_guess_type = 'ai' THEN 1 ELSE 0 END) AS false_negative
-            FROM user_guesses
-            JOIN images ON user_guesses.image_id = images.image_id
-        """)
-
-        # Execute the query and commit the transaction
-        result = db.session.execute(query)
-        db.session.commit()
-
-        # Get the results as a dictionary
-        row = result.fetchone()
-        if not row:
-            return {"error": "No data found"}
-
-        true_positive = row[0]
-        false_positive = row[1]
-        true_negative = row[2]
-        false_negative = row[3]
-
-        # Calculate the metrics
-        accuracy = (true_positive + true_negative) / (true_positive + false_positive + true_negative + false_negative)
-        precision = true_positive / (true_positive + false_positive) if (true_positive + false_positive) != 0 else 0
-        recall = true_positive / (true_positive + false_negative) if (true_positive + false_negative) != 0 else 0
-        f1_score = 2 * (precision * recall) / (precision + recall) if (precision + recall) != 0 else 0
-
-        return {
-            "accuracy": accuracy,
-            "precision": precision,
-            "recall": recall,
-            "f1Score": f1_score
-        }
-
-    except Exception as e:
-        db.session.rollback()
-        return {"error": str(e)}
 
 def get_feedback_instances():
     try:
@@ -305,33 +134,7 @@ def get_feedback_resolution_status():
         db.session.rollback()
         return {"error": str(e)}
 
-def get_matching_feedback_for_image(image_id):
-    try:
-        query = text(f"""
-            SELECT 
-                feedback.feedback_id, 
-                feedback.msg, 
-                feedback.x AS x, 
-                feedback.y AS y
-            FROM feedback
-            JOIN feedback_users ON feedback_users.feedback_id = feedback.feedback_id
-            JOIN user_guesses ON feedback_users.guess_id = user_guesses.guess_id
-            JOIN images ON user_guesses.image_id = images.image_id
-            WHERE user_guesses.image_id = {image_id}
-            AND user_guesses.user_guess_type = images.image_type;
-        """)
-        result = db.session.execute(query)
-        db.session.commit()
 
-        rows = []
-        for row in result:
-            row_dict = {column: value for column, value in zip(result.keys(), row)}
-            rows.append(row_dict)
-
-        return rows
-    except Exception as e:
-        db.session.rollback()
-        return {"error": str(e)}
 
 def get_random_unresolved_feedback(image_id):
     try:
@@ -359,201 +162,14 @@ def get_random_unresolved_feedback(image_id):
         return {"error": str(e)}
 
 
-def get_feedback_with_filters(image_type=None, resolved=None, sort_by=None, sort_order='asc', limit=20, offset=0):
-    try:
-        # Start the query string
-        query_str = """
-            SELECT 
-                images.image_id,
-                images.image_path,
-                images.image_type,
-                COUNT(CASE WHEN feedback.resolved IS false THEN 1 END) AS unresolved_count,
-                MAX(feedback.date_added) AS last_feedback_time,
-                images.upload_time
-            FROM images
-            LEFT JOIN user_guesses ON user_guesses.image_id = images.image_id
-            LEFT JOIN feedback_users ON feedback_users.guess_id = user_guesses.guess_id
-            LEFT JOIN feedback ON feedback.feedback_id = feedback_users.feedback_id
-
-            WHERE 1=1
-        """
-
-        # Apply filters for image_type if provided
-        if image_type and image_type != "all":
-            query_str += " AND images.image_type = :image_type"
-
-        query_str += """
-            GROUP BY images.image_id, images.image_path, images.image_type, images.upload_time
-        """
-
-        # Apply HAVING clause if resolved is False
-        if resolved is False:
-            query_str += """
-                HAVING COUNT(CASE WHEN feedback.resolved IS false THEN 1 END) > 0
-            """
-        if resolved:
-            query_str += """
-                HAVING COUNT(CASE WHEN feedback.resolved IS false THEN 1 END) = 0
-"""
-
-        # Apply sorting based on sort_by field
-        valid_sort_fields = ['last_feedback_time', 'unresolved_count', 'upload_time']
-        if sort_by:
-            if sort_by in valid_sort_fields:
-                query_str += f" ORDER BY {sort_by} {sort_order.upper()}"
-            elif sort_by == "image_id":
-                query_str += f" ORDER BY images.image_id {sort_order.upper()}"
-            else:
-                raise ValueError("Invalid sort field provided.")
-        else:
-            query_str += f" ORDER BY last_feedback_time {sort_order.upper()}"
-
-        # Add LIMIT and OFFSET for pagination
-        query_str += f" LIMIT :limit OFFSET :offset"
-
-        # Prepare parameters for query execution
-        params = {
-            'limit': limit,
-            'offset': offset
-        }
-
-        # If image_type filter is provided, add to params
-        if image_type:
-            params['image_type'] = image_type
-        if resolved is not None:
-            params['resolved'] = resolved
-
-        # Execute the query
-        query = text(query_str)
-        result = db.session.execute(query, params)
-
-        # Process and format the results
-        feedback_data = []
-        for row in result.mappings():
-            feedback_data.append({
-                'image_id': row['image_id'],
-                'image_path': row['image_path'],
-                'image_type': row['image_type'],
-                'unresolved_count': row['unresolved_count'],
-                'last_feedback_time': (
-                    row['last_feedback_time'].strftime('%Y-%m-%d')
-                    if isinstance(row['last_feedback_time'], datetime) else row['last_feedback_time']
-                ),
-                'upload_time': (
-                    row['upload_time'].strftime('%Y-%m-%d')
-                    if isinstance(row['upload_time'], datetime) else row['upload_time']
-                ),
-            })
-
-        return feedback_data
-
-    except Exception as e:
-        print(f"Error fetching feedback: {e}")
-        return []
-
-def get_feedback_count(image_type=None, resolved=None):
-    try:
-        count_query_str = """
-            SELECT COUNT(DISTINCT images.image_id) 
-            FROM images
-            LEFT JOIN user_guesses ON user_guesses.image_id = images.image_id
-            LEFT JOIN feedback_users ON feedback_users.guess_id = user_guesses.guess_id
-            LEFT JOIN feedback ON feedback.feedback_id = feedback_users.feedback_id
-            WHERE 1=1
-        """
-
-        if image_type and image_type != "all":
-            count_query_str += " AND images.image_type = :image_type"
-
-        if resolved is not None:
-            count_query_str += f" AND feedback.resolved IS {'true' if resolved else 'false'}"
-
-        count_query = text(count_query_str)
-        params = {
-            'image_type': image_type,
-            'resolved': resolved
-        }
-        count_result = db.session.execute(count_query, params)
-
-        total_count = count_result.scalar()
-
-        return total_count
-
-    except Exception as e:
-        print(f"Error fetching feedback count: {e}")
-        return 0
 
     
 
-def get_image_by_id(image_id):
-    try:
-        query = text("""
-            SELECT 
-                image_id, 
-                image_path, 
-                image_type, 
-                upload_time 
-            FROM images 
-            WHERE image_id = :image_id
-        """)
-        result = db.session.execute(query, {"image_id": image_id})
-        db.session.commit()
-
-        row = result.fetchone()
-        
-        if row:
-            image_data = {column: value for column, value in zip(result.keys(), row)}
-            return image_data
-        else:
-            return None
-
-    except Exception as e:
-        db.session.rollback()
-        print(f"Error fetching image by id: {e}")
-        return None
-
-def fetch_data_for_csv(table_name):
-    try:
-        query = text(f"SELECT * FROM {table_name}")
-        
-        result = db.session.execute(query)
-        db.session.commit()
-
-        columns = result.keys()
-
-        rows = result.fetchall()
-
-        data = [dict(zip(columns, row)) for row in rows]
-
-        return data
-
-    except Exception as e:
-        db.session.rollback()
-        raise Exception(f"Error fetching data for table {table_name}: {e}")
-
-def get_metadata_counts():
-    try:
-        queries = {
-            'feedback': "SELECT COUNT(*) FROM feedback_users",
-            'image': "SELECT COUNT(*) FROM images",
-            'leaderboard': "SELECT COUNT(*) FROM user_guesses",
-            'competition': "SELECT COUNT(*) FROM competitions" 
-        }
-
-        counts = {}
 
 
-        for table, query in queries.items():
-            result = db.session.execute(text(query))
-            db.session.commit()
-            count = result.fetchone()[0]
-            counts[table] = count
 
-        return counts
 
-    except Exception as e:
-        db.session.rollback()
-        return {"error": str(e)}
+
 
 UPLOAD_FOLDER = '../MedGenAI-Images/Images/'
 ALLOWED_EXTENSIONS = {'png', 'jpg', 'jpeg', 'gif'}
@@ -628,40 +244,7 @@ def upload_image_service(request, image_type):
     return jsonify({'error': 'Invalid file format'}), 400
 
 
-def resolve_all_feedback_by_image(image_id: int):
-    try:
-        query = text("""
-            SELECT ug.guess_id
-            FROM user_guesses ug
-            JOIN images img ON ug.image_id = img.image_id
-            WHERE img.image_id = :image_id
-        """)
-        result = db.session.execute(query, {'image_id': image_id})
-        db.session.commit()
 
-        guess_ids = [row[0] for row in result]
-
-        if not guess_ids:
-            return {"error": "No guesses found for the given image_id"}
-
-        update_query = text("""
-            UPDATE feedback
-            SET resolved = TRUE
-            WHERE feedback_id IN (
-                SELECT fu.feedback_id
-                FROM feedback_users fu
-                JOIN feedback f ON fu.feedback_id = f.feedback_id
-                WHERE fu.guess_id IN :guess_ids
-            )
-        """)
-        db.session.execute(update_query, {'guess_ids': tuple(guess_ids)})
-        db.session.commit()
-
-        return {"message": "All feedback for the image has been marked as resolved"}
-
-    except Exception as e:
-        db.session.rollback()
-        return {"error": str(e)}
 
 def filter_users_by_tags(tag_names, match_all=True, sort_by="level", desc=True):
     """
